refactor: extract helper function
