+++ libs/bite-tribe/api/src/lib/utils/user-collection-key.ts
@@ -1,0 +1,1 @@
+export const USERS_COLLECTION = 'users';
